==> scripts/S03_individual_fly.Rmd <==
---
title: "individual_fly_analysis"
author: "daisy"
date: "2025-08-18"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(here)
source(here::here("scripts", "S00_data_source.R"))
```

## Data importing

```{r block1, echo=FALSE}
#Importing data 
library(readxl)
individual_fly_data <- read_excel(
  path = here::here("data", "individual_fly_data.xlsx")
  )
```

## Data plotting
To make the order of the four genotypes and the three treatment consistent.
```{r block2, echo=FALSE}
#Changing order of genotype
unique(individual_fly_data$genotype)
individual_fly_data$genotype <- factor(
  individual_fly_data$genotype,
  levels = c("tT", "mT", "tM", "mM")
  )

#Changing order of treatment
unique(individual_fly_data$treatment)
individual_fly_data$treatment <- factor(
  individual_fly_data$treatment,
  levels = c("constant", "fluctuation_small", "fluctuation_big")
  )

```

Plotting a boxplot:
```{r block3, echo=FALSE}
library(ggplot2)

ggplot(individual_fly_data, aes(x=treatment, y=dev_time, fill=genotype)) +
  geom_boxplot(position = position_dodge(width = 0.8)) +
  scale_fill_manual(values = colour_palette) + 
  xlab("Genotype at Different Treatments") +
  ylab("Developmental Time (Hours)") +
  ggtitle("Developmental Time by Genotype and Treatments")
```

```{r}
ggsave(
  "devtime.png",
  plot = last_plot(),
  path = here::here("results"),
  width = 10,
  height = 7
)
```

Those for the fluctuation big are across much bigger range is due to the fact that many of the flies from high fluctuation tends to emerge on a Friday, followed by a Monday count after the weekend.

#Calculating mean and sd bars

```{r block10, echo=FALSE}
summary_data <- individual_fly_data %>%
  group_by(treatment, genotype) %>%
  summarise(
    mean_dev = mean(dev_time, na.rm = TRUE),
    sd_dev   = sd(dev_time, na.rm = TRUE),
  )
```

Trying to put some sd bars on the plot:
```{r block8, echo=FALSE}
ggplot(summary_data, aes(x = treatment, y = mean_dev, colour = genotype)) +
  geom_point(position = position_dodge(width = 0.6), size = 3) +
  geom_errorbar(
    aes(ymin = mean_dev - sd_dev, ymax = mean_dev + sd_dev),
    width = 0.2,
    position = position_dodge(width = 0.6)
  ) +
  scale_color_manual(
    values = colour_palette)+
  xlab("Genotype at Different Treatments") +
  ylab("Developmental Time (Hours)") +
  ggtitle("Mean Developmental Time (±SD) by Genotype and Treatment") +
  theme_minimal() +
  theme(
    panel.grid.major = element_line(colour ="grey", linetype = "dotdash"),
    plot.title = element_text(hjust = 0.6, face = "bold", size = 14)
  )
```

```{r}
ggsave(
  "devtime_mean2.png",
  plot = last_plot(),
  path = here::here("results"),
  width = 10,
  height = 7
)
```

*Per condition:
The SD bars overlap quite a lot - which mean in terms of the spread of data, there might not be a huge difference across genotypes

However, the means are different (and are significantly different evidenced by the TukeyHSD below)

#Trying violin plot instead
This potentially allow better show of the distributions + give more information to the fluctuation_big (with a much larger range than the other two treatments)

```{r block9, echo=FALSE}
library(ggplot2)

ggplot(individual_fly_data, aes(x = treatment, y = dev_time, fill = genotype)) +
  geom_violin(position = position_dodge(width = 0.8), trim = FALSE) + 
  
  scale_fill_manual(values = colour_palette) +
  xlab("Genotype at Different Treatments") +
  ylab("Developmental Time (Hours)") +
  ggtitle("Developmental Time by Genotype and Treatments") +
  theme_minimal()

```

```{r block9, echo=FALSE}
library(ggplot2)

ggplot(individual_fly_data, aes(x = treatment, y = dev_time, fill = genotype)) +
  geom_violin(position = position_dodge(width = 0.8), trim = FALSE) +  # Violin plot
  stat_summary(
    fun = mean,
    fun.min = function(x) mean(x) - sd(x),
    fun.max = function(x) mean(x) + sd(x),
    geom = "errorbar",
    width = 0.2,
    position = position_dodge(width = 0.8),
    color = "black"
  ) +
  stat_summary(
    fun = mean,
    geom = "point",
    position = position_dodge(width = 0.8),
    color = "black",
    size = 2
  ) +
  scale_fill_manual(values = colour_palette) +
  xlab("Genotype at Different Treatments") +
  ylab("Developmental Time (Hours)") +
  ggtitle("Developmental Time by Genotype and Treatments") +
  theme_minimal()

```
