==> scripts/S04_survival.Rmd <==
---
title: "survival_data"
author: "daisy"
date: "2025-08-19"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(here)
source(here::here("scripts", "S00_data_source.R"))
```

#Data importing

```{r block1, echo=FALSE}
#Importing data 
library(readxl)
survival_data <- read_excel(
  path = here::here("data", "survival_data.xlsx")
  )
```

#Survival rate
Calculating survival rates by dividing number of fly count (emerge) per vial by 20 (expected total number of flies = number of eggs present in each vial)
```{r block2, echo=FALSE}
survival_data <- survival_data %>%
  mutate(survival = adults / 20) 
```

#Survival score
Due to individual errors when conducting the experiments, e.g. where more than 20 eggs are accidentally put into vials, we transfer the survival rates into "survival scores". The survival scores refines survival rates > 1 to just 1.00 (100%). This is allowed due to considering the fact that if more than 20 flies are emerged per vial, the vial should already shown evidence of (very) high survival.
```{r block2, echo=FALSE}
survival_data <- survival_data %>%
  mutate(survival_score = pmin(survival, 1)) 
```

#Export a copy of the dataset to results (calculated survival scores)
```{r blcok3, echo=FALSE}
library(writexl)
write_xlsx(survival_data, "survival_results.xlsx")
```

#Plotting
Plotting a boxplot:
```{r block4, echo=FALSE}
#Changing order of genotype
unique(survival_data$genotype)
survival_data$genotype <- factor(
  survival_data$genotype,
  levels = c("tT", "mT", "tM", "mM")
  )

#Changing order of treatment
unique(survival_data$treatment)
survival_data$treatment <- factor(
  survival_data$treatment,
  levels = c("constant", "fluctuation_small", "fluctuation_big")
  )

```


```{r block5, echo=FALSE}
library(ggplot2)

ggplot(survival_data, aes(x=treatment, y=survival_score, fill=genotype)) +
  geom_boxplot(position = position_dodge(width = 0.8)) +
  scale_fill_manual(values = colour_palette) + 
  xlab("Genotype at Different Treatments") +
  ylab("Survival Score") +
  ggtitle("Survival Score by Genotype and Treatments")

```

Adding mean and sd bars:
```{r block10, echo=FALSE}
library(ggplot2)

ggplot(survival_data, aes(x = treatment, y = survival_score, fill = genotype)) +
  geom_boxplot(position = position_dodge(width = 0.8)) +
  stat_summary(
    fun = mean,
    fun.min = function(x) mean(x) - sd(x),
    fun.max = function(x) mean(x) + sd(x),
    geom = "errorbar",
    width = 0.2,
    position = position_dodge(width = 0.8),
    color = "black"
  ) +
  stat_summary(
    fun = mean,
    geom = "point",
    position = position_dodge(width = 0.8),
    color = "black",
    size = 2
  ) +
  scale_fill_manual(values = colour_palette) +
  xlab("Genotype at Different Treatments") +
  ylab("Survival Score") +
  ggtitle("Survival Score by Genotype and Treatments")


```

#Calculating mean and sd bars
```{r block6, echo=FALSE}
summary_data <- survival_data %>%
  group_by(treatment, genotype) %>%
  summarise(
    mean_survival = mean(survival_score, na.rm = TRUE),
    sd_survival   = sd(survival_score, na.rm = TRUE),
  )
```

#Plotting only mean and sd bars
```{r block7, echo=FALSE}
library(ggplot2)

ggplot(summary_data, aes(x=treatment, y=mean_survival, colour=genotype)) +
geom_point(position = position_dodge(width = 0.6), size = 3) +
  geom_errorbar(
    aes(ymin = mean_survival - sd_survival, ymax = mean_survival + sd_survival),
    width = 0.2,
    position = position_dodge(width = 0.6)
  ) +
  scale_colour_manual(values = colour_palette) + 
  xlab("Genotype at Different Treatments") +
  ylab("Survival Score") +
  ggtitle("Mean Survival Score (±SD) by Genotype and Treatments") + 
  theme_minimal() +
  theme(
    plot.background = element_rect(fill="grey95"),
    panel.grid.major = element_line(colour ="grey95"),
    plot.title = element_text(hjust = 0.6, face = "bold", size = 14)
  )

```

```{r}
ggsave(
  "survival_mean.png",
  plot = last_plot(),
  path = here::here("results"),
  width = 10,
  height = 7
)
```
